facade class implemented. refactored. support for display done, display float. some changes in data processor API
