+++ DataProcessor/DataProcessor.h
@@ -11,7 +11,7 @@
 #include <iterator>
 #include <assert.h>
 #include "Commons.h"
-#include "InternalStorage.h"
+#include "Storing.h"
 
 using namespace std;
 class DataProcessor{
@@ -28,7 +28,7 @@ public:
 	string addTask(Data task);
 	string displayTask(TimeMacro startTime, TimeMacro endTime);
 	string deleteTask(int number);
-	string clearTask(TimeMacro startTime, TimeMacro endTime);
+	string clearTask();
 	string executeUndo();
 	string searchTask(string keyword);
 	string editTask(int taskNumber, Data task);
