+++ Main/heading_template.txt
@@ -1,0 +1,1 @@
+uCode	macroTBeg		macroTEnd		microTBeg	microTEnd	completeStatus	priority	alarmMacro		alarmMicro	desc
